[CELEBORN-2176] Fix uncaught exception in DataPusher
### What changes were proposed in this pull request?
If DataPusher throws exception not like  CelebornIOException or IOException, do not just log it in ThreadExceptionHandler but set exception ref to avoid data lost.

### Why are the changes needed?
As described in jira, such problem will make pushthread stopped and lead to data lost.
The problem job which occurs this issue
<img width="600" height="400" alt="image" src="https://github.com/user-attachments/assets/6eedbf35-4ef7-4785-ae2b-7714ed86b93b" />

and the normal job which no exeception throws during pushthread working
<img width="600" height="400" alt="image" src="https://github.com/user-attachments/assets/c818da8b-ab58-4bf2-85c0-a7e4da958398" />

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
by UT

Closes #3507 from buska88/celeborn-2176.

Authored-by: lijianfu03 <lijianfu03@meituan.com>
Signed-off-by: SteNicholas <programgeek@163.com>