[CELEBORN-2166] Fast fail reduce stage if shuffle data is lost because of worker lost
### What changes were proposed in this pull request?

- Fix the WorkerStatusTracker logic, so unknown workers are marked correctly in excluded workers.
- Trigger shuffle data lost if the worker hosting the shuffle data is lost.

This can be extended to –
- fast fail mapper stages as well before the commit starts.
- with push replicate enabled with multiple workers loss.

### Why are the changes needed?

Currently even if worker crashs or became unavailable for some reason and marked as lost by Master, reduce stage still try to read data from it and fail after running for sometime which is in-efficient. We can detect this early and fail the reduce stage with SHUFFLE_DATA_LOST before starting the stage.

### Does this PR introduce _any_ user-facing change?

NA

### How was this patch tested?
WIP

Closes #3496 from s0nskar/CELEBORN-2166.

Authored-by: Sanskar Modi <sanskarmodi97@gmail.com>
Signed-off-by: SteNicholas <programgeek@163.com>