[CELEBORN-2194] Change default value of celeborn.worker.directMemoryRatioForReadBuffer
### What changes were proposed in this pull request?

Change default value of `celeborn.worker.directMemoryRatioForReadBuffer` from 0.1 to 0.35.

### Why are the changes needed?

The default value of `celeborn.worker.directMemoryRatioForReadBuffer` is 0.1, which is too small to cause a backlog of read buffer requests in `ReadBufferDispacther`. Therefore, `celeborn.worker.directMemoryRatioForReadBuffer` should be changed from `0.1` to `0.35` which is production practice value to raise read buffer threshold of `ReadBufferDispatcher`.

### Does this PR resolve a correctness bug?

No.

### Does this PR introduce _any_ user-facing change?

The default value of `celeborn.worker.directMemoryRatioForReadBuffer` is changed to 0.35.

### How was this patch tested?

CI.

Closes #3527 from SteNicholas/CELEBORN-2194.

Authored-by: SteNicholas <programgeek@163.com>
Signed-off-by: SteNicholas <programgeek@163.com>