[CELEBORN-2217] Use a separate thread to handle RpcEndpointVerifier messages
### What changes were proposed in this pull request?

An EndpointVerifierMessageLoop is introduced to specifically handle RpcEndpointVerifier messages, and a separate thread is used to execute the EndpointVerifierMessageLoop.

### Why are the changes needed?

RpcEndpointVerifier.CheckExistence is a high-priority message that needs to be processed first. In the original model, if the LifecycleManager's RPC message queue accumulates a large backlog, RpcEndpointVerifier.CheckExistence messages cannot be processed immediately, leading to numerous ShuffleClient initialization failures and causing task failures.

### Does this PR resolve a correctness bug?

No

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Cluster Test

Closes #3554 from TheodoreLx/verifier-first.

Authored-by: TheodoreLx <1548069580@qq.com>
Signed-off-by: Shuang <lvshuang.xjs@alibaba-inc.com>